+++ prank.h
@@ -94,20 +94,20 @@ namespace ukb {
 	  typename std::vector<vertex_descriptor>::iterator end = V.end();
 	  float norm = 0.0;
 	  for (; v != end; ++v) {
-		double rank = 0.0;
+		float rank=0.0;
 		typename graph_traits<G>::in_edge_iterator e, e_end;
 		tie(e, e_end) = in_edges(*v, g);
 		for(; e != e_end; ++e) {
 		  vertex_descriptor u = source(*e, g);
 		  if (*v == u) continue; // No self-loops
-		  rank += rank_map1[u] * wmap[*e] * static_cast<double>(out_coef[u]);
+		  rank += rank_map1[u] * wmap[*e] * out_coef[u];
 		}
-		double dangling_factor = 0.0;
+		float dangling_factor = 0.0;
 		if (0.0 == out_coef[*v]) {
 		  // dangling link
-		  dangling_factor = static_cast<double>(damping)*rank_map1[*v];
+		  dangling_factor = damping*rank_map1[*v];
 		}
-		rank_map2[*v] = damping*rank + (dangling_factor + static_cast<double>(1.0 - damping) )*ppv_V[*v];
+		rank_map2[*v] = damping*rank + (dangling_factor + 1.0 - damping )*ppv_V[*v];
 		norm += fabs(rank_map2[*v] - rank_map1[*v]);
 	  }
 	  return norm;
@@ -135,7 +135,7 @@ namespace ukb {
 	  float damping = 0.85;
 	  // Initialize rank_map1 appropriately
 	  {
-		const double init_value = 1.0f/static_cast<double>(V.size());
+		const float init_value = 1.0f/static_cast<float>(V.size());
 		typename graph_traits<G>::vertex_iterator v, end;
 		for (tie(v, end) = vertices(g); v != end; ++v) {
 		  rank_map1[*v]=init_value;
